Added the following functionalities for android and tested:
- getState() -> idle, recording, paused
- recordHDVideo()
- isBusyRecording()

isBusyRecording returns a boolean if media projection manager or hb recorder is already in recording, but getState returns the state of the recording initialized by the hb recorder itself upon startRecording called.

these functions have been added to the ExpoScreenRecorderModule.kt and implemented according to the recording logic. However, these are commented and marked TODO in src/index.ts because logically their implementation is correct, but they don't work as intended. I doubt it's a HB Recorder issue. I will create an issue on their git page and try to fix these in the next pull request.

 - pauseRecording()
 - resumeRecording()